switch to logging instead
